requirements.txt
Flask
Requests
BeautifulSoup4
sentence-transformers
torch
bs4
newspaper3k

